Add support for listing Stale issues.
In this commit we enhance the tool to have ability to list stale issues.
The issue being stale is decided based on date last updated.
By default the date, older than which a issue will be considered stale
is 30 days. Also we can specify the date to consider issues stale back
from as well we can have issues filtered based on labels defined in Repo.